+++ CHANGELOG.md
@@ -135,7 +135,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
   pipelines.
 
 [unreleased]: https://github.com/recipe/secretsweeper/compare/0.1.1...HEAD
-[0.1.0]: https://github.com/recipe/secretsweeper/compare/0.1.0...0.1.1
+[0.1.1]: https://github.com/recipe/secretsweeper/compare/0.1.0...0.1.1
 [0.1.0]: https://github.com/recipe/secretsweeper/compare/0.0.1-alpha.8...0.1.0
 [0.0.1-alpha.8]: https://github.com/recipe/secretsweeper/compare/0.0.1-alpha.7...0.0.1-alpha.8
 [0.0.1-alpha.7]: https://github.com/recipe/secretsweeper/compare/0.0.1-alpha.6...0.0.1-alpha.7